Require a version bump to be initially releasable.

+++ lib/shoe.rb
@@ -100,7 +100,7 @@ class Shoe
       puts spec.to_ruby
     end
 
-    if there_is_no_tag_for_the_current_version && we_are_on_the_master_branch && we_have_already_pushed_the_master_branch_to_a_remote_called_origin
+    if the_current_version_is_not_zero && there_is_no_tag_for_the_current_version && we_are_on_the_master_branch && we_have_already_pushed_the_master_branch_to_a_remote_called_origin
       desc "Release #{spec.name}-#{spec.version}"
       task :release do
         File.open("#{spec.name}.gemspec", 'w') { |f| f.write spec.to_ruby }
@@ -124,6 +124,10 @@ class Shoe
 
   def everything_in_the_bin_directory
     File.directory?('bin') ? Dir.entries('bin') - ['.', '..'] : []
+  end
+
+  def the_current_version_is_not_zero
+    spec.version != '0.0.0'
   end
 
   def there_are_any_work_in_progress_features
